feat: progressInterval clamped from 0.1 to 5

+++ android/src/main/java/com/amazonaws/ivs/reactnative/player/AmazonIvsViewManager.kt
@@ -184,7 +184,7 @@ class AmazonIvsViewManager : SimpleViewManager<AmazonIvsView>(),
 
   override fun setProgressInterval(
     view: AmazonIvsView?,
-    value: Int
+    value: Double
   ) {
     view?.setProgressInterval(value)
   }
